fix: ui fixed and code cleanup
