+++ src/Interfaces/ILanguage.ts
@@ -7,4 +7,6 @@ export interface ILanguage {
 
 export interface ILanguageListProps {
   languages: ILanguage[];
+  navigationCallback?: (path: string) => void;
+  type: string;
 }
