+++ src/components/LanguageChangeModal/LanguageChangeModal.tsx
@@ -1,25 +1,24 @@
 /* eslint-disable no-param-reassign */
 import React, { useLayoutEffect, useState } from "react";
-import { useTranslation } from "react-i18next";
 import { useRecoilState, useRecoilValue } from "recoil";
 
 import { Modal } from "antd";
 
-import { vibrateWorks } from "@src/constants/vibrateCheck";
 import { ILanguage } from "@src/Interfaces";
 import { i18n } from "@src/translations/i18n";
 
 import { themeState } from "@src/store/ThemeState";
 import { languageChangeModal, languageSelectionState, darkModeState } from "@src/store";
 import "./index.scss";
+import { getLanguages } from "@src/utils";
+import { LanguagesList } from "@components/LanguageChoice/LanguagesList";
 
 export const LanguageChangeModal = () => {
-  const { t } = useTranslation();
   const [, setPosition] = useState(1);
   const open = useRecoilValue(languageChangeModal);
   const darkModeStatus = useRecoilValue(darkModeState);
   const theme = useRecoilValue(themeState);
-  const [IsLanguageChosen, setIsLanguageChosen] = useRecoilState(languageSelectionState);
+  const [IsLanguageChosen] = useRecoilState(languageSelectionState);
   const langSelected = (lang: string, newPos: number) => {
     if (i18n.language.includes(lang)) {
       setPosition(newPos);
@@ -28,70 +27,21 @@ export const LanguageChangeModal = () => {
     return false;
   };
 
-  let languages: ILanguage[] = [
-    {
-      sno: 1,
-      title: t("english"),
-      langId: "en",
-      selected: true,
-    },
-    {
-      sno: 2,
-      title: t("french"),
-      langId: "fr",
-      selected: false,
-    },
-    {
-      sno: 3,
-      title: t("dutch"),
-      langId: "nl",
-      selected: false,
-    },
-    {
-      sno: 4,
-      title: t("hindi"),
-      langId: "hi",
-      selected: false,
-    },
-    {
-      sno: 5,
-      title: t("spanish"),
-      langId: "es",
-      selected: false,
-    },
-    {
-      sno: 6,
-      title: t("german"),
-      langId: "de",
-      selected: false,
-    },
-    {
-      sno: 7,
-      title: t("portuguese"),
-      langId: "pt",
-      selected: false,
-    },
-  ];
+  const languages: ILanguage[] = getLanguages();
 
   const [Languages, setLanguages] = useState(languages);
 
   useLayoutEffect(() => {
     languages.forEach((element) => {
       if (langSelected(element.langId, element.sno)) {
+        languages[0].sno = element.sno;
         languages[0].selected = false;
+        element.sno = 1;
         element.selected = true;
       }
     });
-    languages = languages.sort((a, b) => a.sno - b.sno);
     setLanguages(languages);
   }, [IsLanguageChosen]);
-
-  const handleClick = (langId: string) => {
-    vibrateWorks ? navigator.vibrate(100) : null;
-    setIsLanguageChosen(langId);
-    i18n.changeLanguage(langId);
-    localStorage.setItem("language", JSON.stringify(langId));
-  };
 
   return (
     <Modal
@@ -105,21 +55,7 @@ export const LanguageChangeModal = () => {
         darkModeStatus ? "dark" : "light"
       }-theme${theme[darkModeStatus ? "dark" : "light"]}`}
     >
-      <div className="languageListWrapper">
-        {Languages.map((lang: ILanguage) => (
-          <button
-            key={String(lang.sno)}
-            type="button"
-            className={`languageListButton ${lang.selected} ? "selected" : ""`}
-            onClick={() => {
-              handleClick(lang.langId);
-            }}
-          >
-            <input style={{ marginRight: "10px" }} type="radio" checked={lang.selected} readOnly />
-            {lang.title}
-          </button>
-        ))}
-      </div>
+      <LanguagesList languages={Languages} type="modal" />
     </Modal>
   );
 };
