+++ src/components/LanguageChoice/LanguagesList.tsx
@@ -1,23 +1,21 @@
 import React from "react";
 import i18n from "i18next";
 
+import { useRecoilState } from "recoil";
 import { ILanguageListProps, ILanguage } from "@src/Interfaces/ILanguage";
-import { useNavigate } from "react-router-dom";
 import { vibrateWorks } from "@src/constants/vibrateCheck";
 import { languageSelectionState } from "@src/store";
-import { useRecoilState } from "recoil";
 
 export const LanguagesList = (props: ILanguageListProps) => {
-  const { languages } = props;
-  const navigate = useNavigate();
+  const { languages, navigationCallback, type } = props;
   const [, setIsLanguageChosen] = useRecoilState(languageSelectionState);
 
   const handleClick = (langId: string) => {
-    (vibrateWorks) ? navigator.vibrate(100) : null;
+    vibrateWorks ? navigator.vibrate(100) : null;
     setIsLanguageChosen(langId);
     i18n.changeLanguage(langId);
     localStorage.setItem("language", JSON.stringify(langId));
-    navigate("/ZinZenFAQ");
+    if (type === "fragment" && navigationCallback) navigationCallback("/ZinZenFAQ");
   };
   return (
     <div className="containerLang">
@@ -26,7 +24,9 @@ export const LanguagesList = (props: ILanguageListProps) => {
           key={String(lang.sno)}
           type="button"
           className={lang.selected ? "selected" : ""}
-          onClick={() => { handleClick(lang.langId); }}
+          onClick={() => {
+            handleClick(lang.langId);
+          }}
         >
           {lang.title}
           <input type="radio" checked={lang.selected} readOnly />
