+++ src/pages/LandingPage/LandingPage.scss
@@ -2,7 +2,7 @@
   font-size: 0.875rem;
   font-weight: 500;
   color: #000 !important;
-  margin-top: 16px !important;
+  margin-block: 16px !important;
 }
 
 .containerLang {
@@ -14,7 +14,6 @@
   background: transparent;
   border-radius: 0.5rem;
   box-shadow: 0px 5px 6px rgba(0, 0, 0, 0.25);
-  margin-top: 14px;
 
   button {
     width: 100%;
