+++ src/utils/index.ts
@@ -1,4 +1,6 @@
+import { useTranslation } from "react-i18next";
 import { GoalItem } from "@src/models/GoalItem";
+import { ILanguage } from "@src/Interfaces";
 import { languagesFullForms } from "@src/translations/i18n";
 
 export async function createContactRequest(url: string, body: object | null = null, method = "POST") {
@@ -215,4 +217,54 @@ export const getMonthByIndex = (index: number) => {
     "December",
   ];
   return month[index];
+};
+
+export const getLanguages = () => {
+  const { t } = useTranslation();
+  const languages: ILanguage[] = [
+    {
+      sno: 1,
+      title: t("english"),
+      langId: "en",
+      selected: true,
+    },
+    {
+      sno: 2,
+      title: t("french"),
+      langId: "fr",
+      selected: false,
+    },
+    {
+      sno: 3,
+      title: t("dutch"),
+      langId: "nl",
+      selected: false,
+    },
+    {
+      sno: 4,
+      title: t("hindi"),
+      langId: "hi",
+      selected: false,
+    },
+    {
+      sno: 5,
+      title: t("spanish"),
+      langId: "es",
+      selected: false,
+    },
+    {
+      sno: 6,
+      title: t("german"),
+      langId: "de",
+      selected: false,
+    },
+    {
+      sno: 7,
+      title: t("portuguese"),
+      langId: "pt",
+      selected: false,
+    },
+  ];
+
+  return languages;
 };
